+++ crates/ludtwig/src/check/rule.rs
@@ -121,22 +121,14 @@ impl CheckResult {
         self
     }
 
-    pub fn secodary_note<S: Into<String>>(mut self, syntax_range: TextRange, message: S) -> Self {
+    /// The secondary (blue) label which can provide more context and explain the error to a user.
+    pub fn secondary_note<S: Into<String>>(mut self, syntax_range: TextRange, message: S) -> Self {
         self.secondary.push(CheckNote {
             syntax_range,
             message: message.into(),
         });
         self
     }
-
-    // The secondary (blue) label which can provide more context and explain the error to a user.
-    // pub fn secondary_note<S: Into<String>>(mut self, syntax_range: TextRange, message: S) -> Self {
-    //     self.secondary.push(CheckNote {
-    //         syntax_range,
-    //         message: message.into(),
-    //     });
-    //     self
-    // }
 
     /// Add a code suggestion which the user can follow or is replaced automatically
     pub fn suggestion<R: Into<String>, S: Into<String>>(